Module rustbook::collatz

collatz.rs : test for the Collatz conjecture

HISTORY: The Collatz conjecture is a conjecture in mathematics that concerns a sequence defined as follows:

start with any positive integer n. Then each term is obtained from the previous term as follows:

if the previous term is even, the next term is one half of the previous term.

If the previous term is odd, the next term is 3 times the previous term plus 1.

The conjecture is that no matter what value of n, the sequence will always reach 1.

The conjecture is named after Lothar Collatz, who introduced the idea in 1937, two years after receiving his doctorate.

See source code on GitHub

Functions

collatz
Collatz function

main
Collatz example

reverse_collatz
Collatz : reverse problem