Function rustbook::guessing_game::run

pub fn run()

let - statements are used to create variables

expect - Unwraps a results, yielding the content of an Ok. Panics if the value is an Err.

variables are immutable by default. Adding mut in front of name turns it into a mutable variable.

String::new() - Indicates new() is an associated function with the type String. Associated functions are implemented on a type (static function)

& - Indicates a reference to a variable. References are immutable by default. Need to add &mut guess to make mutable.

Ordering - Enum with variants {Less, Greater, Equal}

trim() - Eliminates white space and newlines. read_line(&var) will assign the value given by the use AND \n.

loop - Creates an infinite loop.

===== General =====

When calling methods with '.', push to new line.

Run cargo doc --open to build documentation of all dependancies and open in browser.

guess.cmp(&secret_number) cannot infer that guess should be converted to a numberical type for comparison.