Function rustbook::ownership::what_is_ownership::run

pub fn run()

Ownership Rules

Each value in Rust has a variable that's called owner.

There can only be one owner at a time.

When the owner groes out of scope, the value will be dropped.

String

len - How much memory in bytes the String is currently using

capacity - Total memory allocated by OS to String

General

Two types of strings: string literals "...", and String

When variables on the heap go out of scop, drop() is called to free memory.

When pointers are copied from one variable to another, the first variable is invalidated. This is considered a move. This only applies to variables on the stack.

Rust has a Copy trait that can by placed on anything stored on the stack. Anything with the Drop trait cannot implement the Copy trait.

Returning values can also transfer ownership